M1/S1 - Added Welcome View, created reusable components, settled VIPER module for Welcome
