+++ MovieDBFullProject/Common/Constants.swift
@@ -1,0 +1,8 @@
+//
+//  Constants.swift
+//  MovieDBFullProject
+//
+//  Created by Cesar Humberto Grifaldo Garcia on 04/10/23.
+//
+
+import Foundation
