+++ MovieDBFullProject/Extensions/UIFont+NewFonts.swift
@@ -1,0 +1,8 @@
+//
+//  UIFont+NewFonts.swift
+//  MovieDBFullProject
+//
+//  Created by Cesar Humberto Grifaldo Garcia on 04/10/23.
+//
+
+import Foundation
