+++ MovieDBFullProject/Protocols/Onboarding/OnboardingProtocols.swift
@@ -1,0 +1,8 @@
+//
+//  OnboardingProtocols.swift
+//  MovieDBFullProject
+//
+//  Created by Cesar Humberto Grifaldo Garcia on 04/10/23.
+//
+
+import Foundation
