+++ MovieDBFullProject/Resources/Colors/UIColor+NewColors.swift
@@ -1,0 +1,8 @@
+//
+//  Colors.swift
+//  MovieDBFullProject
+//
+//  Created by Cesar Humberto Grifaldo Garcia on 04/10/23.
+//
+
+import Foundation
