+++ MovieDBFullProject/UIComponents/UIButtons/CommonUIButton.swift
@@ -1,0 +1,20 @@
+//
+//  CustomUIButton.swift
+//  MovieDBFullProject
+//
+//  Created by Cesar Humberto Grifaldo Garcia on 13/10/23.
+//
+
+import UIKit
+
+class CustomUIButton: UIButton {
+
+    /*
+    // Only override draw() if you perform custom drawing.
+    // An empty implementation adversely affects performance during animation.
+    override func draw(_ rect: CGRect) {
+        // Drawing code
+    }
+    */
+
+}
