+++ MovieDBFullProject/UIComponents/UIViewControllers/BaseViewController.swift
@@ -1,0 +1,29 @@
+//
+//  BaseViewController.swift
+//  MovieDBFullProject
+//
+//  Created by Cesar Humberto Grifaldo Garcia on 04/10/23.
+//
+
+import UIKit
+
+class BaseViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Do any additional setup after loading the view.
+    }
+    
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destination.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
