--- MovieDBFullProject/ViewController.swift
@@ -1,19 +1,0 @@
-//
-//  ViewController.swift
-//  MovieDBFullProject
-//
-//  Created by Cesar Humberto Grifaldo Garcia on 04/10/23.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
